# apps/web/src/lib/api/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "AddProviderModelRequest": {
        "properties": {
          "context_window_tokens": {
            "minimum": 1.0,
            "title": "Context Window Tokens",
            "type": "integer"
          },
          "display_name": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Display Name",
            "type": "string"
          },
          "kind": {
            "pattern": "^(chat|embedding)$",
            "title": "Kind",
            "type": "string"
          },
          "model_name": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Model Name",
            "type": "string"
          },
          "supports_structured_output": {
            "default": false,
            "title": "Supports Structured Output",
            "type": "boolean"
          },
          "supports_tools": {
            "default": false,
            "title": "Supports Tools",
            "type": "boolean"
          },
          "supports_vision": {
            "default": false,
            "title": "Supports Vision",
            "type": "boolean"
          }
        },
        "required": [
          "model_name",
          "display_name",
          "kind",
          "context_window_tokens"
        ],
        "title": "AddProviderModelRequest",
        "type": "object"
      },
...
      "CreateProviderRequest": {
        "properties": {
          "base_url": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url"
          },
          "credential_plaintext": {
            "maxLength": 4096,
            "minLength": 1,
            "title": "Credential Plaintext",
            "type": "string"
          },
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "provider_type": {
            "pattern": "^(openai|anthropic|gemini|ollama)$",
            "title": "Provider Type",
            "type": "string"
          }
        },
        "required": [
          "provider_type",
          "name",
          "credential_plaintext"
        ],
        "title": "CreateProviderRequest",
        "type": "object"
      },
...
      "ErrorDetailResponse": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetailResponse",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/ErrorDetailResponse"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ModelResponse": {
        "properties": {
          "context_window_tokens": {
            "title": "Context Window Tokens",
            "type": "integer"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "model_name": {
            "title": "Model Name",
            "type": "string"
          },
          "price_per_1m_input_tokens": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Per 1M Input Tokens"
          },
          "price_per_1m_output_tokens": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Per 1M Output Tokens"
          },
          "provider_id": {
            "format": "uuid",
            "title": "Provider Id",
            "type": "string"
          },
          "provider_name": {
            "title": "Provider Name",
            "type": "string"
          },
          "provider_type": {
            "title": "Provider Type",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "supports_structured_output": {
            "title": "Supports Structured Output",
            "type": "boolean"
          },
          "supports_tools": {
            "title": "Supports Tools",
            "type": "boolean"
          },
          "supports_vision": {
            "title": "Supports Vision",
            "type": "boolean"
          },
          "team_id": {
            "format": "uuid",
            "title": "Team Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "team_id",
          "provider_id",
          "provider_type",
          "provider_name",
          "model_name",
          "display_name",
          "kind",
          "status",
          "context_window_tokens",
          "supports_tools",
          "supports_vision",
          "supports_structured_output",
          "price_per_1m_input_tokens",
          "price_per_1m_output_tokens"
        ],
        "title": "ModelResponse",
        "type": "object"
      },
      "ProviderResponse": {
        "properties": {
          "base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "provider_type": {
            "title": "Provider Type",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "team_id": {
            "format": "uuid",
            "title": "Team Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "team_id",
          "provider_type",
          "name",
          "base_url",
          "status"
        ],
        "title": "ProviderResponse",
        "type": "object"
      },
...
      "SetProviderModelPricingRequest": {
        "properties": {
          "price_per_1m_input_tokens": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Per 1M Input Tokens"
          },
          "price_per_1m_output_tokens": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Per 1M Output Tokens"
          }
        },
        "title": "SetProviderModelPricingRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Hify API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/providers": {
      "post": {
        "operationId": "providers_create_provider",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProviderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Permission denied"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Provider",
        "tags": [
          "providers"
        ]
      }
    },
    "/providers/models": {
      "get": {
        "operationId": "providers_list_provider_models",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ModelResponse"
                  },
                  "title": "Response Providers List Provider Models",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Permission denied"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          }
        },
        "summary": "List Provider Models",
        "tags": [
          "providers"
        ]
      }
    },
    "/providers/models/{model_id}/pricing": {
      "put": {
        "operationId": "providers_set_provider_model_pricing",
        "parameters": [
          {
            "in": "path",
            "name": "model_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Model Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetProviderModelPricingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Permission denied"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Provider Model Pricing",
        "tags": [
          "providers"
        ]
      }
    },
    "/providers/{provider_id}/models": {
      "post": {
        "operationId": "providers_add_provider_model",
        "parameters": [
          {
            "in": "path",
            "name": "provider_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Provider Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddProviderModelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Permission denied"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Provider Model",
        "tags": [
          "providers"
        ]
      }
    },
...
  }
}